options_monitor_agent/requirements.txt
# === Core ===
anthropic>=0.30.0
python-dotenv>=1.0.0

# === Data ===
yfinance>=0.2.31
pandas>=2.1.0
numpy>=1.26.0

# === Visualization ===
matplotlib>=3.8.0
plotly>=5.18.0
rich>=13.6.0

# === Greeks ===
scipy>=1.11.0

# === Database ===
sqlalchemy>=2.0.0

# === Scheduling ===
schedule>=1.2.0

# === Notifications ===
python-telegram-bot>=20.6
aiohttp>=3.9.0

# === Dashboard ===
flask>=3.0.0
flask-socketio>=5.3.0

# === Backtesting ===
tabulate>=0.9.0
